Add files via upload
This code is NON FUNCTIONAL currently. Its more of a proof of concept and a learning experience for me. I wanted to
experience what real world Aerospace Programmers might use python for. 

In theory: This code should Read telemetry data out of a CSV file -> Process the data from the file including 
time, position and velocity data. -> Create a visualization of the trajectory using matplotlib