Add files via upload
This is a music bot designed for Discord. It has functionality for various commands. !join to join the voice chat you're in. !play <URL> to play a song from YouTube, !stop and !leave. I've uploaded it without my Bot token for safety. Importing the modules and getting them all working in the environment took way longer than expected. It turns out I just had to restart my computer because I was doing it on windows and I edited PATH. -Facepalm-